Improvements and fixes
