+++ SwrveSDK/src/main/java/com/swrve/sdk/messaging/SwrveInAppCampaign.java
@@ -10,6 +10,7 @@ import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import java.io.File;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Date;
@@ -44,7 +45,7 @@ public class SwrveInAppCampaign extends SwrveBaseCampaign {
             JSONArray jsonMessages = campaignData.getJSONArray("messages");
             for (int k = 0, t = jsonMessages.length(); k < t; k++) {
                 JSONObject messageData = jsonMessages.getJSONObject(k);
-                SwrveMessage message = createMessage(campaignManager, this, messageData);
+                SwrveMessage message = createMessage(this, messageData, campaignManager.getCacheDir());
 
                 // If the message has some format
                 if (message.getFormats().size() > 0) {
@@ -163,8 +164,8 @@ public class SwrveInAppCampaign extends SwrveBaseCampaign {
         return null;
     }
 
-    protected SwrveMessage createMessage(ISwrveCampaignManager campaignManager, SwrveInAppCampaign swrveCampaign, JSONObject messageData) throws JSONException {
-        return new SwrveMessage(swrveCampaign, messageData, campaignManager);
+    protected SwrveMessage createMessage(SwrveInAppCampaign swrveCampaign, JSONObject messageData, File cacheDir) throws JSONException {
+        return new SwrveMessage(swrveCampaign, messageData, cacheDir);
     }
 
     /**
